% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_oc_plot.R
\name{generate_oc_plot}
\alias{generate_oc_plot}
\title{Generate the plot(s) displaying operating characteristics for a single or multiple designs}
\usage{
generate_oc_plot(`3+3` = NULL,
                        `BOIN` = NULL,
                        `i3+3` = NULL)
}
\arguments{
\item{3+3}{the object returned by run.sim.3()}

\item{BOIN}{the object returned by run.sim.b()}

\item{i3+3}{the object returned by run.sim.i3()}
}
\value{
\code{generate_oc_plot()} returns figures displaying the operating characteristics for the user-specified design(s).
}
\description{
Generate the plot(s) displaying operating characteristics for a single or multiple designs
}
