% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.true.R
\name{load.true}
\alias{load.true}
\title{Load several commonly used scenarios}
\usage{
load.true(pT = 0.25)
}
\arguments{
\item{pT}{the target DLT rate.}
}
\value{
\code{load.true()} returns:

(1) a matrix (\code{$p.true}) containing the true toxicity probabilities of the investigational dose levels.

(2) a vector (\code{$mtd.true}) which specifies the true MTD.
}
\description{
Load several commonly used scenarios
}
