% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path}
\alias{path}
\title{Finds Local path}
\usage{
path(path = NULL)
}
\arguments{
\item{path}{Optional string input.}
}
\value{
String
}
\description{
Queries working directory and returns tha base directory as a string.
}
