% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/this.R
\name{this}
\alias{this}
\title{Finds Local path}
\usage{
this(path = NULL)
}
\arguments{
\item{path}{Optional string input.}
}
\value{
String
}
\description{
Queries working directory and returns the base user dir.
}
