% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_decision_table.R
\name{generate_decision_table}
\alias{generate_decision_table}
\title{Generate the decision table(s) for a single or multiple designs}
\usage{
generate_decision_table(`3+3` = NULL,
                               `BOIN` = NULL,
                               `mTPI2` = NULL,
                               `i3+3` = NULL,
                               `G3` = NULL)
}
\arguments{
\item{3+3}{the object returned by get.decision.3()}

\item{BOIN}{the object returned by get.decision.boin()}

\item{mTPI2}{the object returned by get.decision.mtpi2()}

\item{i3+3}{the object returned by get.decision.i3()}

\item{G3}{the object returned by get.decision.g3()}
}
\value{
\code{generate_decision_table()} returns a figure displaying the decision table(s) for the user-specified design(s).
}
\description{
Generate the decision table(s) for a single or multiple designs
}
