% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_oc_plot.R
\name{generate_oc_plot}
\alias{generate_oc_plot}
\title{Generate the plot(s) displaying operating characteristics for a single or multiple designs}
\usage{
generate_oc_plot(`3+3` = NULL,
                        `BOIN` = NULL,
                        `mTPI2` = NULL,
                        `i3+3` = NULL,
                        `G3` = NULL)
}
\arguments{
\item{3+3}{the object returned by run_sim_threethree()}

\item{BOIN}{the object returned by run_sim_boin()}

\item{mTPI2}{the object returned by run_sim_mtpi2()}

\item{i3+3}{the object returned by run_sim_i3()}

\item{G3}{the object returned by run_sim_g3()}
}
\value{
\code{generate_oc_plot()} returns figures displaying the operating characteristics for the user-specified design(s).
}
\description{
Generate the plot(s) displaying operating characteristics for a single or multiple designs
}
